+++ apps/c/shsgc/CUDA/fact_kernel_cuda_kernel.cu
@@ -102,20 +102,20 @@ void ops_par_loop_fact_kernel_execute(ops_kernel_descriptor *desc) {
   dim3 grid( (x_size-1)/block->instance->OPS_block_size_x+ 1, 1, 1);
   dim3 tblock(block->instance->OPS_block_size_x,1,1);
 
-
-
-  int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size);
-  int dat1 = (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size);
+  long long int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size
+                                                 : args[0].dat->elem_size);
+  long long int dat1 = (block->instance->OPS_soa ? args[1].dat->type_size
+                                                 : args[1].dat->elem_size);
 
   char *p_a[2];
 
   //set up initial pointers
-  int base0 = args[0].dat->base_offset + 
-           dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
+  long long int base0 = args[0].dat->base_offset +
+                        dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
   p_a[0] = (char *)args[0].data_d + base0;
 
-  int base1 = args[1].dat->base_offset + 
-           dat1 * 1 * (start[0] * args[1].stencil->stride[0]);
+  long long int base1 = args[1].dat->base_offset +
+                        dat1 * 1 * (start[0] * args[1].stencil->stride[0]);
   p_a[1] = (char *)args[1].data_d + base1;
